Coverity fixes
* MOPPAND-713

+++ configuration-lib/src/main/java/ee/ria/DigiDoc/configuration/util/FileUtils.java
@@ -1,16 +1,15 @@
 package ee.ria.DigiDoc.configuration.util;
 
 import java.io.BufferedReader;
-import java.io.BufferedWriter;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
+import java.nio.charset.StandardCharsets;
 
 import timber.log.Timber;
 
@@ -25,7 +24,7 @@ public class FileUtils {
     }
 
     public static String readFileContent(InputStream inputStream) {
-        try (BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream))) {
+        try (BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream, StandardCharsets.UTF_8))) {
             StringBuilder sb = new StringBuilder();
             int i;
             while((i = reader.read()) != -1) {
@@ -60,8 +59,12 @@ public class FileUtils {
 
     public static void storeFile(String filePath, String content) {
         File file = new File(filePath);
-        file.getParentFile().mkdirs();
-        try (BufferedWriter writer = new BufferedWriter(new FileWriter(file))) {
+        boolean isDirsCreated = file.getParentFile().mkdirs();
+        if (isDirsCreated) {
+            Timber.d("Directories created for %s", filePath);
+        }
+        try (FileOutputStream fileStream = new FileOutputStream(file.getAbsoluteFile());
+             OutputStreamWriter writer = new OutputStreamWriter(fileStream, StandardCharsets.UTF_8)) {
             writer.write(content);
         } catch (IOException e) {
             throw new IllegalStateException("Failed to store file '" + filePath + "'!", e);
@@ -70,7 +73,10 @@ public class FileUtils {
 
     public static void storeFile(String filePath, byte[] content) {
         File file = new File(filePath);
-        file.getParentFile().mkdirs();
+        boolean isDirsCreated = file.getParentFile().mkdirs();
+        if (isDirsCreated) {
+            Timber.d("Directories created for %s", filePath);
+        }
         try (FileOutputStream os = new FileOutputStream(file)) {
             os.write(content);
         } catch (IOException e) {
@@ -81,7 +87,10 @@ public class FileUtils {
     public static void createDirectoryIfNotExist(String directory) {
         File destinationDirectory = new File(directory);
         if (!destinationDirectory.exists()) {
-            destinationDirectory.mkdirs();
+            boolean isDirsCreated = destinationDirectory.mkdirs();
+            if (isDirsCreated) {
+                Timber.d("Directories created for %s", directory);
+            }
         }
     }
 
@@ -90,12 +99,17 @@ public class FileUtils {
     }
 
     public static void removeFile(String filePath) {
-        new File(filePath).delete();
+        if (fileExists(filePath)) {
+            boolean isFileDeleted = new File(filePath).delete();
+            if (isFileDeleted) {
+                Timber.d("File %s deleted", filePath);
+            }
+        }
     }
 
     public static void writeToFile(BufferedReader reader, String destinationPath, String fileName) {
         try (FileOutputStream fileStream = new FileOutputStream(new File(destinationPath + File.separator + fileName));
-             OutputStreamWriter writer = new OutputStreamWriter(fileStream)) {
+             OutputStreamWriter writer = new OutputStreamWriter(fileStream, StandardCharsets.UTF_8)) {
 
             String fileLine;
             while ((fileLine = reader.readLine()) != null) {
